Update patch set 1
Patch Set 1:

I tried rewording this a bit, to include that, but it got too wordy.  I think it's pretty clear how this is meant to be used as is.  The less than 1,000,000,000 is the major point to clarify, since how we check that timespecs are valid.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0