Update patch set 2
Patch Set 2:

The kernel does save and restore except x0 and x1 (return values) and SP (for the clone case) at entry.S. However, I am not sure if this is an implementation detail from Linux, and thus susceptible to change over the releases; or an ABI de-facto that follows the Linux rules of not breaking userspace. I am checking with an Linux arm64 maintainer.

Patch-set: 2
Reviewer: Gerrit User 1331125 <1331125@85c56323-6fa9-3386-8a01-6480fb634889>