Update patch set 5
Patch Set 5: Code-Review+2

(1 comment)

By the way, this is good enough for me to add caching for my unwinder when I replace the _Unwind_XXX functions. I've been trying to think of a way to cache without this and nothing was 100% so this is a nice addition.

Patch-set: 5
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2