Update patch set 1
Patch Set 1:

unfortunately frameworks/native/libs/binder/ndk/stability.cpp:31 has gone in since:

frameworks/native/libs/binder/ndk/stability.cpp:31:2: error: libbinder_ndk should only be built in a system context
#error libbinder_ndk should only be built in a system context
^


i guess `-target aarch64-linux-android10000` is setting it to 10000, so we can't revert this exactly, we'll need to say "if it's not set or it's set but equal to 10000". does this mean we should _remove_ the L47-L53 part of this file? if the platform build system is now setting this for us?

Patch-set: 1
CC: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0