Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> As things stand, I'm concerned about how late the prctl() is called. Indeed, any call to malloc() shouldn't be done before this point, and stack tagging should be disabled. I have a change locally that issues the prctl() much earlier (both in static executables and the linker), but I feel like it might be too invasive (you cannot use the prctl() wrapper until the TLS is set up).

This is already before any calls to malloc(), but yes, this might need to be moved for stack tagging. Evgenii discovered that this point is not early enough for HWASAN stack tagging either (therefore we're relying on the prctl in the HWASAN runtime). I would prefer that we take care of that once we're closer to actually turning stack tagging on.

In terms of the TLS dependency, we could avoid that by issuing the syscall directly (i.e. via assembly), but again that's something to worry about once we get there.

> If the initial exclusion mask were set to all tags but 0, these problems should go away, because tagging operations would be no-ops until the prctl() is issued. I suggested that the default exclusion mask be changed on the list, please reply there if you have any comment: https://lore.kernel.org/linux-arm-kernel/ef61bbc6-76d6-531d-2156-b57efc070da4@arm.com/

Not entirely I think, but defaulting the mask to 0xffff seems like a good idea anyway. I'll reply on that thread with more detail.

Patch-set: 3
Reviewer: Gerrit User 1067098 <1067098@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0