Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review-2
> 
> i don't think i understand what we're trying to do here... libc is sufficiently low level that only children of the zygote can use this. so adding an API that doesn't actually work in most cases seems more misleading than helpful. we'll need the public APIs for stuff like MTE, so why not just have the code to deal with features and experiments live in the zygote, where it will always work, and won't mislead folks into thinking they can guard their changes in libc generally in a way that always works.

And gating on app compat changes shouldn't make sense on non zygote process.
AFAIU this is exactly as targetSDK is currently working in libc, this is just a bit more power and precision. We can change the name because it is a bit general and perhaps misleading.
I've seen already 2 app affecting changes that wanted this information in libc, and there may be more - do you think it's better for each one of those to pipe things into libc? 
This seems like a way to be consistent across changes and make it easy to use compat framework.

Patch-set: 6
Reviewer: Gerrit User 1463094 <1463094@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0