Update patch set 3
Patch Set 3:

> I looked at the ELF libraries before and after this change, and I noticed that the order of the RUNPATH entries changes. Typically, RUNPATH has $ORIGIN and some other directory, and these two components are swapped. For each library, I verified that $ORIGIN and the other directory never have the same library name.

Yeah, this is an artifact of how we're appending / prepending variables in the Android.mk/bp files:

Android.build.testlib.internal.mk used appending:

  # 2. Set dt_runpath to origin to resolve dependencies
  $(module)_ldflags += -Wl,--rpath,\$${ORIGIN} -Wl,--enable-new-dtags

While the equivalent in the Android.bp is using the bionic_testlib_defaults defaults module, and defaults are prepended to the individual module's flags.

And for the host, this also gets a bit weird because Soong adds more entries to the rpath for host tests by default. In make it was just $ORIGIN/../lib64 and $ORIGIN/lib64, and in Soong, for host tests it also adds $ORIGIN/../../lib64, $ORIGIN/../../../lib64, and $ORIGIN/ (so we've got both $ORIGIN/ and ${ORIGIN} for some of these)

Patch-set: 3
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0