Update patch set 1
Patch Set 1:

Ping for "This is a port of https://android-review.googlesource.com/c/platform/bionic/+/1254251.  Are we planning to make a similar change to bionic in aosp/master as well?"

If not, we should prioritize b/147765083 before the next toolchain update.

Patch-set: 1
Reviewer: Gerrit User 1058860 <1058860@85c56323-6fa9-3386-8a01-6480fb634889>