math.h: match other OSes more closely.
isnan() and isinf() are type-generic macros, not usually declared as functions. Remove the function declarations and the big comment talking about how we shouldn't expose them.

isnanl() and isinfl() are exposed as functions by glibc (but not the BSDs), and we've always had the definitions, so we can increase source compatibility there trivially.

See the discussion on https://github.com/llvm/llvm-project/pull/96008 for more background.

Change-Id: I798153f710794870bb4864f2ae8b98147f9d8763