math_test.cpp: simplify and extend the generic macro captures.
This code hadn't been updated since stlport days (judging by the comment), and libc++ plays the same trick with all the <math.h> type-generic macros. Use a helper macro to remove the duplicate bodies of all the functions, and use #define to avoid having to rewrite tests to explicitly use the macro-capturing-functions (or even remember/know that they exist).

Change-Id: I7c38f28bb33398ee44a51f6e5dd832cc51c491c2