Remove gwp_asan_crash_handler dependency from libc
This was a single function that we actually needed
(__gwp_asan_error_is_mine) that I've just moved into 'gwp_asan' as well.
This allows us to remove the dependency on 'gwp_asan_crash_handler',
which originally resulted in duplicate definitions of other stuff that
was used in both libraries (revealed in aosp/3150976).

Bug: N/A
Test: m libc
Change-Id: Ibf1a0a5abb3a58f9b06bd43731f079df5503f5c7