Make LD_DEBUG logging go to stderr as well as logcat.
This matches glibc, and is a lot more convenient. Although in general
bionic tries to send diagnostics to both, I'm not entirely convinced
that sending this to logcat makes any sense; it's too verbose to be
useful there, and it takes so long that there's bound to be other output
interleaved. But "do what we do elsewhere" fixes my immediate usability
problem, and adding behavior seems less contentious than changing
behavior.

Change-Id: I5d6c17b5c35a75c19a890cb357984c729b1a4b52