linker: map large portion of ELF file to read its fragments
ElfReader::Read() maps ELF file fragments separately to read their
data, which leads to at least 4 separate mmap() syscalls and additional
mmap() call for every PT_NOTE section. This can be replaced with one
mmap() syscall which covers large enough address space to include all
required areas in majority of cases and use the old approach for a few
outliers. Local testing on a Pixel 7 device reveals that 1MB is enough
to cover 97.5% of the cases, therefore we map the first 1MB of the Elf
file (which starts at file_offset_). pread64() syscall in ReadElfHeader()
can also be eliminated since it's always at the beginning of the Elf
file and will be included in this large mapping.
This reduction in syscalls should benefit app launch time since
ElfReader::Read() is used by load_library(), which is one of the heavily
used functions during application launch.

Bug: 322132947
Test: atest -c linker-unit-tests
Change-Id: I7b25151aba52da4116c2b6afd8a254646d0b622c
Signed-off-by: Suren Baghdasaryan <surenb@google.com>