Don't manually -DFLT_EVAL_METHOD.
FLT_EVAL_METHOD comes from <float.h> and is how the compiler communicates with us, not the other way around.

Also, the FreeBSD source #includes <float.h> in all the files that actually need the STRICT_ASSIGN() macro that uses FLT_EVAL_METHOD, so we don't need the forced include now (if we ever did).

Change-Id: I12d2867c6ca382f2849c8fa0adb5078f95792fbf