Simplify *.map.txt for API levels <= 21.
There's no need to record that history; we're not doing so in the headers any more, and it makes this file harder to read.

In particular, it makes the absence of sigblock() and sigsetmask() on riscv64 stick out like sore thumbs. Yes, they're useless, but, yes, it's also useless to go to extra effort to make riscv64 different from arm64/x86-64. I'll come back to clean that up.

Change-Id: Ic24430d0f9e83c1915214dfb6d3e1190e4bafc8a