Use size_t for stdio's __sbuf even on ILP32.
This is probably more of a theoretical bug than a practical one, since allocating enough memory to get yourself into trouble in a 32-bit process is non-trivial, but there's no obvious reason for this LP64/ILP32 difference here.

At the time we ported bionic to LP64, this struct was still in <stdio.h>, which probably explains the reticence to just change ILP32 from int to size_t. Even though those types have the same size and alignment, and it's unlikely that anyone's code would be broken by a change of signedness, the same change changed the short that contains an fd in __sFILE to int, which definitely is an ABI change (that at the time at least was app-breaking, because apps were definitely poking about in __sFILE directly).

Change-Id: I0b938fb6cb0da3a241a5ba4a4cda1b867a2578c1