Use ELF aliases for most of the strto*_l/wcsto*_l functions.
We don't need any of these no-op implementations that the compiler turns into a single branch instruction.

There is some __INTRODUCED_IN cleanup we can do here, but I'm going to do that separately as part of a broader "let's just __RENAME any of the functions that take [but ignore] a locale_t to the function that does the real work" header-only change, where we can do more.

Change-Id: I208768f95d5554455b755d21e8eb2c8a21000101