Re-land Fix dlext tests for MTE globals
Existing dlext anonymous tests use /proc/maps to figure out which
mappings to create a duplicate of, and then branch to them and ensure
the symbol resolution ends up in the right place.

Unfortunately, this breaks with MTE globals (and seems slightly broken
already with things like bss?) because the mapping name doesn't exactly
match. The global variable sections, under MTE globals, are remapped as
PROT_MTE which requires them to be anonymous, and so we use the kernel
API to name them with a mt: prefix. This prefix trips up the
/proc/maps selection.

Instead of just implementing lazy-matching, I've changed the logic to
instead use dl_iterate_phdr to ensure we grab all the sections from the
DSO for now and in future. This seems more bullet-proof.

Test: atest bionic-unit-tests on a fullmte device with MTE globals
re-landed, enabled, and the new toolchain. Also on a non-MTE device.
Bug: 320448268

Change-Id: Ib4a89382d15cba2d3b50f62329b2d64ed927de8e