Remove guards on constants and types.
The only remaining guard is the one on the posix_madvise constants,
because that's one where we had a bug filed for breaking some 3p code
where the presence of those macros was being used as a proxy for
runtime availability of posix_madvise. That's wrong, but it's in the
wild so we can't do much about it. That wrong code is still in
upstream boost:
https://github.com/boostorg/interprocess/blob/9fc4ee7dd7277d86c691921007856aca99c3c1ac/include/boost/interprocess/mapped_region.hpp#L811-L814

Bug: https://github.com/android/ndk/issues/1652
Test: treehugger
Change-Id: If5762422524e1667014b4b4ce6a2a2700dd4aff0