Fix posix_spawn()/posix_spawnp() nullability annotations.
argv is terminated by a null pointer, so that was clearly a mistake. Linux actually allows argv itself to be a null pointer, though that's not portable and even Linux calls pr_warn_once() if you take advantage of this, so we might want to remove that _Nullable from all the exec()-like functions, but for now let's just make sure everything's consistent and not _wrong_... We can worry about not enabling bad practices later (or not, since this isn't something I've ever seen anyone do).

Change-Id: I633245085ca960e6e7b73d4a8aeb4c0535034cef