16kb: bionic: Re-align libtest_invalid-textrels.so to 16kb
The library libtest_invalid-textrels.so was realigned to 16kb
for arm64 and x86_64.

In order for the bionic-unit-tests to pass with 16kb page size targets,
it is necessary to realign the libraries to 16kb.

This library was generated using:

```
$ m geninvalelf64
$ adb push out/target/product/vsoc_x86_64_pgagnostic/system/bin/geninvalelf64 /data
$ adb shell
  vsoc_x86_64_pgagnostic:./data/geninvalelf64 test_X86_64/libtest_empty.so test_X86_64
```

With this change, the test `dlopen_invalid_textrels` is passing in
aosp_cf_x86_64_phone_pgagnostic target.

```
$ m bionic-unit-tests
$ atest bionic-unit-tests
[2870/2974] dlfcn#dlopen_invalid_textrels: PASSED (2ms)
```

Bug: 347728628
Test: Use readelf to verify the dynamic entries
For arm64:

```
 $ readelf -d arm64/libtest_invalid-textrels.so | grep TEXTREL
 0x000000000000001e (FLAGS)              TEXTREL BIND_NOW
```

For x86_64

```
 $ readelf -d x86_64/libtest_invalid-textrels.so | grep TEXTREL
 0x000000000000001e (FLAGS)              TEXTREL BIND_NOW
```

Change-Id: I6f3e29ca4fe6c3ebe8f82809a6cb6a519126a3e8