16kb: bionic: Re-align libtest_invalid-textrels2.so to 16kb
The library libtest_invalid-textrels2.so was realigned to 16kb
for arm64 and x86_64.

In order for the bionic-unit-tests to pass with 16kb page size targets,
it is necessary to realign the libraries to 16kb.

This library was generated using:

```
$ m geninvalelf64
$ adb push out/target/product/vsoc_x86_64_pgagnostic/system/bin/geninvalelf64 /data
$ adb shell
  vsoc_x86_64_pgagnostic:./data/geninvalelf64 test_X86_64/libtest_empty.so test_X86_64
```

With this change, the test `dlopen_invalid_textrels2` is passing in
aosp_cf_x86_64_phone_pgagnostic target.

```
$ m bionic-unit-tests
$ atest bionic-unit-tests
[2871/2974] dlfcn#dlopen_invalid_textrels2: PASSED (3ms)
```

Bug: 347728628
Test: Use readelf to verify the dynamic entries
For arm64:

```
 $ readelf -d arm64/libtest_invalid-textrels2.so  | grep TEXTREL
 0x0000000000000016 (TEXTREL)            0x8

```

For x86_64

```
 $ readelf -d x86_64/libtest_invalid-textrels2.so  | grep TEXTREL
 0x0000000000000016 (TEXTREL)            0x8
```

Change-Id: I20ba5186d51e57b8f0023c051015d1cfad964305