Add better jemalloc stat logging.
Instead of trying to do all of the logging in bionic, call into jemalloc
and use a callback to log all of the stats for all size classes. The
previous version only dumped the bins, not all of the classes beyond
the bins.

Also, removed the references to huge bytes, this was never enabled
in the current jemalloc.

Test: Ran dumpsys meminfo --logstats and verified log data.
Change-Id: Ieba4096923a31e1216a4b5b8f1f3368c84ec856b