cpp: Explicitly retain array definitions
Some arrays get default-retained when we reach the semicolon, because
the default solution is to treat them as functions. This strips the
semicolon though, leaving us with invalid syntax.

We should really treat arrays (and variable definitions) as "other
declarations", which can be OK to retain, as they don't rely on
kernel-internal data.

This parser is fairly unsound in general, but it isn't much more unsound
to assume that anything with an '=' token at depth 0 will be a variable
or array definition.

NB: there are existing UAPI arrays that *are* retained [1], because they
happen to look like struct declarations. I maintain that behavior as
well, encoded in the tests.

[1] example from asm/amd_hsmp.h:
static const struct hsmp_msg_desc hsmp_msg_desc_table[] = {
...
};

Bug: 369900761
Test: `ANDROID_BUILD_TOP=... pytest ./cpp.py` has no new failures
Test: `bionic/libc/kernel/tools/update_all.py` produces no diff
Change-Id: I5997672436f5ed232fdcd94aafaf792c75fa6bce