Get out of the business of hard-coding a list of uapi unistd.h files.
If we were being clever here (say, excluding the irrelevant x32 or
arm-oabi files) there might be some reason to keep this, but since we're
literally just dragging in "all the things", including the ones we don't
support, let's just let python find them all.

This does actually remove a tiny amount of unwanted cruft from the
generated header as an additional bonus.

Test: ran manually, looked at output
Change-Id: Ie2b349c42d566f3bbe138834febaa0ef2de1c87d