16kb: bionic: Re-align libtest_invalid-local-tls.so to 16kb
The library libtest_invalid-local-tls.so is realigned to 16kb
for arm64 and x86_64.

In order for the bionic-unit-tests to pass with 16kb page size targets,
it is necessary to realign the libraries to 16kb.

This library was generated using the script gen-libtest_invalid-local-tls.sh

With this change, the test `dlopen_invalid_local_tls` is passing in
aosp_cf_x86_64_phone_pgagnostic target.

```
$ m bionic-unit-tests
$ atest bionic-unit-tests
[2876/2978] dlfcn#dlopen_invalid_local_tls: PASSED (3ms)
```

Bug: 347728628
Test: Use readelf to verify alignment
Change-Id: I282e0cb6ee83ad986682cc57bc60eb311ad58f96