Fix races in record allocs.
When recording allocation operations, it's possible to create invalid
traces if operations get out of order because an entry is added
after the operation is completed. To fix this, reserve the entry
before executing the operation, then fill it in afterwards.

Need to allow for a reserved entry to be invalid in cases where the
operation can fail.

Test: All unit tests pass.
Change-Id: Ic7d01a1682da8742f34750e86dd163c085e709a3