Fix an overflow bug in __get_elf_note().
Seen in the wild on ILP32.

There was also a bug where we'd only check the first note in a segment,
which hasn't mattered in practice so far, but seems worth fixing while
we're here (and makes the condition a bit clearer anyway).

Also lift the strlen() out of the loop.

Bug: https://issuetracker.google.com/380356872
Change-Id: I458efb197116f68722ba7fca96fe376fdb6c8352