Remove a historical accident.
It seems like the only reason we weren't maintaining the max alignment on ILP32 was that we weren't supporting huge pages on ILP32. Since it's confusing to have a "max" that's a lie (either as the result of a function call or as member data in ElfReader), and since we're going through all the program headers once anyway, just unconditionally maintain it.

Change-Id: I3e5c22f2f8b455ddbea2297c4bcad28320a995d3