Improve linker workaround documentation for API levels < 23.
Mention SoLoader as well as ReLinker, because SoLoader was probably more popular. Mention the workaround in both of the relevant sections. Make it a bit clearer that this is mostly just for historical interest now (but not quite completely irrelevant, since the NDK still supports back to API 21).

I also note that facebook's SoLoader docs talk about API levels <= 23 instead of API levels < 23 like we do, but without a specific counterexample -- and no-one's ever given us one -- I'm going to assume our docs are correct.

Change-Id: Ic7723028ba8273dfca9fce7834a31baa372a565f