Add fortify memcpy_chk happy path tests.
This catches subtly-broken __memcpy_chk() implementations that fail on
the boundary condition where the requested size is exactly the size of
the destination buffer; this can be validated by changing the x86
__memcpy_chk comparison opcode from `ja` to `jae` and ensuring the test
fails.

Additionally, a test that exercises the non-failure path where the
requested size is less than the buffer size is added, since this wasn't
explicitly tested previously.

Bug: 389669171
Test: ./tests/run-on-host.sh 32 --gtest_filter='*memcpy*'
Test: ./tests/run-on-host.sh 64 --gtest_filter='*memcpy*'
Change-Id: Icd47731f237574b4f8275616137fca8a2d8bb94c