linker: have clang check the DL_ERROR_AFTER() format strings.
The other logging options are checked "for free" by virtue of all macro-expanding to an existing logging function that has the appropriate attribute. This was an exception, and I found and removed a harmless extra argument from one call in my previous change to this code. I expected more fallout from turning on the checking, so kept it for a separate patch, but it seems that the one mistake I already fixed was the only mistake.

Oh well, at least this will mean we can't regress...

Change-Id: I6f47ccf9c545ea5b74375f4a5ceb0951006b948e