__bionic_get_shell_path(): "just work" on the host.
This lets you run _device_ unit tests on the host without setting up
a /system (as run-test.sh does). Before this, isolated gtests just
silently fail.

I initially didn't like the idea of adding an extra system call here,
for fear of breaking seccomp jails ... but then reasoned that if you're
about to shell out (which is the only time this is used: popen(),
system(), execvpe()), you're unlikely to be in a seccomp jail!

Change-Id: I5924daacbfab6524d826355adad5a88bdc11161f